#261: refactor sycn_location test
